Feat: Upgrade to Antlr 4.8
The ethereum/solidity repo is using Antlr4.8. 

I think there should be at least cross testing to see if the output is materially different, less grammar differences from the separate definitions. 

Alternatively, it may be worth considering bundling a separate build output using their grammar and referring to that as the "Strict" parser for consumers to consider relative to their use case.